pyproject.toml
[tool.poetry]
name = "patika-codejam"
version = "0.1.0"
description = ""
authors = ["ahmetveburak <ahmetbozyurtt@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
Django = "^3.2.8"
django-cors-headers = "^3.10.0"
djangorestframework = "^3.12.4"

[tool.poetry.dev-dependencies]
mypy = "^0.910"
black = "^21.9b0"
isort = "^5.9.3"
pre-commit = "^2.15.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 119
exclude = '''
/(
    \.eggs
  | \.git
  | \.mypy_cache
  | \.tox
  | \.venv
  | _build
  | build
  | dist
  | \.migrations
)/
'''

[tool.isort]
line_length = 119
skip_gitignore = true
skip_glob = "*/migrations/*.py"